add: del sights is no events

+++ app/Console/Commands/delDubleType.php
@@ -2,7 +2,9 @@
 
 namespace App\Console\Commands;
 
+use App\Models\Event;
 use App\Models\Sight;
+use App\Models\Status;
 use Illuminate\Console\Command;
 
 class delDubleType extends Command
@@ -32,10 +34,22 @@ class delDubleType extends Command
         $bar = $this->output->createProgressBar(Sight::query()->count());
         Sight::query()->orderBy('id')->chunk(1000, function ($sight) use($bar) {
             $sight->each(function($sight) use($bar) {
-                $types = $sight->types->pluck('id')->toArray();
-                $sight->types()->detach($types);
-                $sight->types()->attach(array_unique($types));
-                $bar->advance();
+                if (Event::where('organization_id', $sight->organization->id)->exists()){
+                    $types = $sight->types->pluck('id')->toArray();
+                    $sight->types()->detach($types);
+                    $sight->types()->attach(array_unique($types));
+                    $bar->advance();
+                } else {
+                    $status = Status::where('name', 'Отказ')->first();
+                    $statuses = $sight->statuses;
+                    foreach($statuses as $status) {
+                        $sight->statuses()->updateExistingPivot($status["id"], [
+                            "last" => false
+                        ]);
+                    }
+                    $sight->statuses()->attach($status->id, ["last" => True]);
+                    $bar->advance();
+                }
             });
         });
         $bar->finish();
